+++ lib/services/runtime/init.ts
@@ -72,6 +72,8 @@ const init = (client: Client, eventHandler: HandleContextEventHandler) => {
       trace: frame,
     });
   });
+
+  return client;
 };
 
 export default init;
